+++ packages/mix_protocol/lib/src/schema/image_styler_codec.dart
@@ -6,7 +6,6 @@ import '../contract/identity_resolution.dart';
 import '../contract/identity_codec.dart';
 import 'common_codecs.dart';
 import 'schema_field.dart';
-import 'styler_field_inventory.dart';
 import 'styler_codec_helpers.dart';
 
 SchemaObject<ImageStyler> imageStylerSchema({
@@ -100,7 +99,6 @@ SchemaObject<ImageStyler> _imageStylerSchemaType(
 
   return stylerSchemaObject<ImageStyler, ImageSpec>(
     rootStyleSchema: rootStyleSchema,
-    ownerFieldInventory: imageStylerInventory,
     fields: [
       image,
       width,